fix(mcp): move profile filter to its own file (line-check: mcp_tools.c >2000)
The P1 additions pushed mcp_tools.c to 2008 lines, tripping the 2000-line
line-check hard limit. mcp_tools.c is chronically at that budget, so move the
profile logic (MCP_CORE_TOOLS + mcp_tool_profile_effective +
mcp_filter_tools_for_profile) into a new mcp_tool_profile.c instead of trimming to
the edge — also a cleaner separation from the tool definitions it filters.

Wired into CORE_SRCS (so the server links it) and the two test link lines in
tests/Rules.mk that reference mcp_tools.o. mcp_tools.c back to 1947 lines;
line-check ok; golden + filter tests pass; aimee-server links clean.
