+++ src/mcp_tools.c
@@ -8,36 +8,7 @@
 #include "headers/session_search_tool.h"
 #include "log.h"
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
-
-/* Tier-0 "core" presentation profile (MCP-native tool names): the high-frequency
- * tools an external MCP client is shown when AIMEE_MCP_TOOL_PROFILE=core|lean.
- * Everything else — including plugin:* and remote-server tools — is hidden from
- * the initial tools/list to shrink the upfront payload. The full catalog stays
- * reachable: P2 adds find_tools/describe_tool for on-demand discovery. Keep this
- * list short and edit it deliberately; it is the floor of what every client sees
- * in lean mode. */
-static const char *const MCP_CORE_TOOLS[] = {
-    "get_help",      "search_docs",                                    /* orient / discover */
-    "search_memory", "memory_recall",     "get_identity",             /* grounding */
-    "find_symbol",   "ast_grep_search",                               /* code intel */
-    "git_status",    "git_commit",        "git_diff_summary",         /* common git */
-    "git_branch",    "git_pr",
-    "delegate",      "ensemble_review",                               /* multi-agent */
-    "ask_user",      "send_message",                                  /* interaction */
-    "create_note",                                                    /* capture */
-    NULL,
-};
-
-static int mcp_name_in_set(const char *name, const char *const *set)
-{
-   for (int i = 0; set[i]; i++)
-      if (strcmp(name, set[i]) == 0)
-         return 1;
-   return 0;
-}
-
 static cJSON *build_tool(const char *name, const char *desc, cJSON *schema)
 {
    cJSON *t = cJSON_CreateObject();
@@ -1973,36 +1944,4 @@ cJSON *mcp_build_tools_list(void)
    cJSON_Delete(remote_tools);
 
    return tools;
-}
-
-const char *mcp_tool_profile_effective(const char *explicit_profile)
-{
-   if (explicit_profile && explicit_profile[0])
-      return explicit_profile;
-   const char *e = getenv("AIMEE_MCP_TOOL_PROFILE");
-   return (e && e[0]) ? e : "full";
-}
-
-int mcp_filter_tools_for_profile(cJSON *tools, const char *profile)
-{
-   if (!tools || !cJSON_IsArray(tools))
-      return 0;
-   profile = mcp_tool_profile_effective(profile);
-   /* "full" (default) presents everything; an unknown profile fails OPEN to the
-    * full set so a typo never silently hides tools. */
-   if (strcmp(profile, "core") != 0 && strcmp(profile, "lean") != 0)
-      return 0;
-
-   int removed = 0;
-   for (int i = cJSON_GetArraySize(tools) - 1; i >= 0; i--)
-   {
-      cJSON *tool = cJSON_GetArrayItem(tools, i);
-      cJSON *nm = cJSON_GetObjectItemCaseSensitive(tool, "name");
-      if (!cJSON_IsString(nm) || !mcp_name_in_set(nm->valuestring, MCP_CORE_TOOLS))
-      {
-         cJSON_DeleteItemFromArray(tools, i);
-         removed++;
-      }
-   }
-   return removed;
 }
